able to get the intial state and one block being emitted, but fail to correctly listen for the next block

+++ src/mempool/domain/fees.rs
@@ -3,9 +3,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Default, Debug, Clone, PartialEq, Serialize, Deserialize)]
 #[serde(rename_all = "camelCase")]
 pub struct Fees {
-    pub fastest_fee: i64,
-    pub half_hour_fee: i64,
-    pub hour_fee: i64,
-    pub economy_fee: i64,
-    pub minimum_fee: i64,
+    pub fastest_fee: f64,
+    pub half_hour_fee: f64,
+    pub hour_fee: f64,
+    pub economy_fee: f64,
+    pub minimum_fee: f64,
 }
