+++ src/mempool/domain/recommended_fees.rs
@@ -4,9 +4,9 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Debug, Clone, Copy)]
 #[serde(rename_all = "camelCase")]
 pub struct RecommendedFees {
-    pub fastest_fee: i64,
-    pub half_hour_fee: i64,
-    pub hour_fee: i64,
-    pub economy_fee: i64,
-    pub minimum_fee: i64,
+    pub fastest_fee: f64,
+    pub half_hour_fee: f64,
+    pub hour_fee: f64,
+    pub economy_fee: f64,
+    pub minimum_fee: f64,
 }
